sw_timer: 32bit int overflow bug fix
Test Plan: No

Reviewers: jwzhang, mingzhou

Reviewed By: mingzhou

Differential Revision: http://10.10.20.8/D1002

+++ dev/soc/arm_cm/le501x/platform.c
@@ -201,7 +201,7 @@ void rco_freq_counting_start()
 
 uint32_t lpcycles_to_hus(uint32_t lpcycles)
 {
-    uint32_t hus = 2*lpcycles*lsi_cnt_val/LSI_CNT_CYCLES;
+    uint32_t hus = 2*(uint64_t)lpcycles*lsi_cnt_val/LSI_CNT_CYCLES;
     //LOG_I("%d,%d",lpcycles,hus);
     return hus;
 }
@@ -229,7 +229,7 @@ uint32_t lsi_freq_update_and_hs_to_lpcycles(int32_t hs_cnt)
         }
     }
     LS_ASSERT(lsi_cnt_val);
-    uint32_t lpcycles = LSI_CNT_CYCLES*625*hs_cnt/2/lsi_cnt_val;
+    uint32_t lpcycles = LSI_CNT_CYCLES*625*(uint64_t)hs_cnt/2/lsi_cnt_val;
     //LOG_I("%d,%d,%d",lsi_cnt_val,lpcycles,hs_cnt);
     return lpcycles;
 }
